Camera Controls
Camera controls modularized out of movement controls.
We don't want to interleave movement with camera controls, especially
with scene specific buttons and such because moving those scripts out of
the scene breaks it.  I might have broken some of the button control
things, but I need to do some other work now so I'll try to fix it soon.
Otherwise I'm sure Cristian can fix it ;)
