+++ ProjectBit3D/ProjectBit/Assets/Scripts/Camera/CameraControl.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: e0978854db61e5145a738a529eb07a26
+timeCreated: 1445800808
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
